#5556: fallback to legacy copy method
