+++ package.json
@@ -55,6 +55,7 @@
     "chrono-node": "^2.6.3",
     "classnames": "^2.3.1",
     "connected-react-router": "^6.9.1",
+    "copy-text-to-clipboard": "^3.1.0",
     "csharp-helpers": "^0.9.3",
     "css-selector-generator": "^3.6.0",
     "date-fns": "^2.28.0",
