+++ src/components/jsonTree/JsonTree.tsx
@@ -48,6 +48,7 @@ import notify from "@/utils/notify";
 import safeJsonStringify from "json-stringify-safe";
 import styles from "./JsonTree.module.scss";
 import AsyncButton from "@/components/AsyncButton";
+import { writeTextToClipboard } from "@/utils/clipboardUtils";
 
 const SEARCH_DEBOUNCE_MS = 100;
 
@@ -158,7 +159,7 @@ const CopyDataButton: React.FunctionComponent<{ data: unknown }> = ({
     aria-label="copy data"
     href="#"
     onClick={async (event) => {
-      await navigator.clipboard.writeText(safeJsonStringify(data, null, 2));
+      await writeTextToClipboard(safeJsonStringify(data, null, 2));
       event.preventDefault();
       event.stopPropagation();
       notify.info("Copied data to the clipboard");
